requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.0.0
chromadb>=0.4.0
Pillow>=8.0.0
requests>=2.26.0

# File format support
openpyxl>=3.0.0  # Excel files
python-docx>=0.8.11  # Word documents
pypdf>=3.0.0  # PDF files (replacing deprecated PyPDF2)
opencv-python>=4.10.0

# Cloud storage support
boto3>=1.26.0  # Amazon S3
google-cloud-storage>=2.0.0  # Google Cloud Storage
azure-storage-blob>=12.0.0  # Azure Blob Storage
minio>=7.2.0  # MinIO Object Storage

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.0.0
flake8>=6.0.0
mypy>=1.0.0
setuptools>=65.0.0
wheel>=0.40.0
setuptools_scm>=7.0.0
pydantic>=2.0.0  # Updated to v2 for Python 3.12 compatibility

# Optional dependencies
tqdm>=4.65.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
botocore~=1.38.27
protobuf~=5.29.5
azure-core~=1.34.0
redis>=4.0.0
Booktype~=1.5
soundfile~=0.13.1

# Testing and documentation
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
myst-parser>=2.0.0